+++ test/Algebra/Graph/Test/AdjacencyMap.hs
@@ -43,8 +43,10 @@ testAdjacencyMap = do
     testDfsForest         t
     testDfsForestFrom     t
     testDfs               t
+    testReachable         t
     testTopSort           t
-    testIsTopSort         t
+    testIsTopSortOf       t
+    testIsAcyclic         t
 
     putStrLn "\n============ AdjacencyMap.scc ============"
     test "scc empty               == empty" $