+++ CHANGES.md
@@ -2,6 +2,9 @@
 
 ## 0.2
 
+* #106: Extend `ToGraph` with algorithms based on adjacency maps.
+* #106: Add `isAcyclic` and `reachable`.
+* #106: Rename `isTopSort` to `isTopSortOf`.
 * #102: Switch the master branch to GHC 8.4.3. Add a CI instance for GHC 8.6.1.
 * #101: Drop `-O2` from the `ghc-options` section of the Cabal file.
 * #100: Rename `fromAdjacencyList` to `stars`.